deps/v8/test/inspector/debugger/wasm-instrumentation-breakpoint-expected.txt

Test instrumentation breakpoints in wasm.

Running test: testBreakInStartFunction
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Compiling wasm module.
Paused at v8://test/compile_module with reason "instrumentation".
Instantiating module.
Paused at v8://test/instantiate with reason "instrumentation".
Paused at wasm://wasm/20da547a with reason "instrumentation".
Script wasm://wasm/20da547a byte offset 26: Wasm opcode 0x01 (kExprNop)
Instantiating a second time (should trigger no breakpoint).
Paused at v8://test/instantiate2 with reason "instrumentation".
Done.

Running test: testBreakInStartFunctionCompileTwice
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Instantiating module.
Paused at v8://test/instantiate with reason "instrumentation".
Paused at wasm://wasm/20da547a with reason "instrumentation".
Script wasm://wasm/20da547a byte offset 26: Wasm opcode 0x01 (kExprNop)
Instantiating a second time (should trigger another breakpoint).
Paused at v8://test/instantiate with reason "instrumentation".
Paused at wasm://wasm/20da547a with reason "instrumentation".
Script wasm://wasm/20da547a byte offset 26: Wasm opcode 0x01 (kExprNop)
Done.

Running test: testBreakInExportedFunction
Setting instrumentation breakpoint
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Instantiating wasm module.
Paused at v8://test/instantiate with reason "instrumentation".
Calling exported function 'func' (should trigger a breakpoint).
Paused at v8://test/call_func with reason "instrumentation".
Paused at wasm://wasm/8c388106 with reason "instrumentation".
Script wasm://wasm/8c388106 byte offset 33: Wasm opcode 0x01 (kExprNop)
Calling exported function 'func' a second time (should trigger no breakpoint).
Paused at v8://test/call_func with reason "instrumentation".
Done.

Running test: testBreakOnlyWithSourceMap
Setting instrumentation breakpoint for source maps only
{
    id : <messageId>
    result : {
        breakpointId : <breakpointId>
    }
}
Instantiating wasm module without source map.
Calling exported function 'func' (should trigger no breakpoint).
Instantiating wasm module with source map.
Calling exported function 'func' (should trigger a breakpoint).
Paused at wasm://wasm/c8e3a856 with reason "instrumentation".
Script wasm://wasm/c8e3a856 byte offset 33: Wasm opcode 0x01 (kExprNop)
Done.
